fix(cli): git setup not creating initial repo
